Fix PSN page scraper typings and module resolution
